+++ app/src/main/java/com/oleksandrlysun/traveminder/receivers/di/ReceiverModule.kt
@@ -1,11 +1,11 @@
-package com.oleksandrlysun.traveminder.services.di
+package com.oleksandrlysun.traveminder.receivers.di
 
-import com.oleksandrlysun.traveminder.services.notification.NotificationPublisher
+import com.oleksandrlysun.traveminder.receivers.notification.NotificationPublisher
 import dagger.Module
 import dagger.android.ContributesAndroidInjector
 
 @Module
-interface ServiceModule {
+interface ReceiverModule {
 
 	@ContributesAndroidInjector
 	fun contributeNotificationPublisherInjector(): NotificationPublisher
